This update has a set of patches from Rui Boon that goes a big step further to seeing the archetype details in a webform.  Thanks Rui.
See: http://localhost:8080/oship/archetypedetails to view the one archetype as far as we have gotten with it.

I have also added back in the atbldr directory.  My (limited) work leads me to believe that my original idea of creating the archetype instances and storing them in the ZODB archetype repository is going to be much easier for new users to create templates for them.  Still work to do on this but I wanted to get Rui's patch commited.
You also need to rerun buildout after this update.  Otherwise you will get an error saying pyparsing not found when you try to start your server.

Cheers,
Tim