Adds the ability to add, edit and display limited fields of archetypes using webforms.
Currently, only schema.TextLine and schema.Bool can be displayed. As such, to see the forms in action, certain temporary adjustments have to be made. Not wanting to pollute the codebase with this, these adjustments have to be manually applied using webformDemo.diff

The addForm is located at http://localhost:8080/oship/@@add and the list of archetype entries stored is at http://localhost:8080/oship/@@arlist

Cheers
Rui Boon